difflib -> string-utils
